==> 10_create_wxdash.Rmd <==
---
title: "WxDash"
resource_files:
- outputs/cwa_estimates/cwa_estimates.dbf
- outputs/cwa_estimates/cwa_estimates.prj
- outputs/cwa_estimates/cwa_estimates.shp
- outputs/cwa_estimates/cwa_estimates.shx
- outputs/county_estimates/county_estimates.dbf
- outputs/county_estimates/county_estimates.prj
- outputs/county_estimates/county_estimates.shp
- outputs/county_estimates/county_estimates.shx
runtime: shiny
output:
  flexdashboard::flex_dashboard:
    orientation: rows
    vertical_layout: fill
    # social: menu
    source_code: "https://github.com/oucrcm/wxdash"
    theme: simplex
---

<head>
<!-- Global site tag (gtag.js) - Google Analytics -->
<script async src="https://www.googletagmanager.com/gtag/js?id=UA-176778534-1"></script>
<script>
  window.dataLayer = window.dataLayer || [];
  function gtag(){dataLayer.push(arguments);}
  gtag('js', new Date());

  gtag('config', 'UA-176778534-1');
</script>
</head> 

```{r global, include = FALSE}
library(leaflet)
library(flexdashboard)
library(leaflet.extras)
library(sf)
library(tidyverse)
library(viridis)
library(miniUI)
library(leaflegend)
library(ggtext)
library(data.table)
library(DT)
library(cowplot)
library(htmlwidgets)
library(srvyr)

vars_to_normalize <- c("TO_SUB_COM", "ALL_READY", "PER_SNO", "TO_RECEP", "PER_FLO", "PER_DRO", "TO_EFFICAC", "HU_SUB_COM",
                       "HU_RESP", "PER_HEA", "PER_TOR", "PER_COL", "TO_RESP", "PER_HUR", "HU_RECEP", "PER_FIR", "HU_OBJ_COM", 
                       "TO_OBJ_COM", "COLD", "FIRE", "FLOOD", "HEAT", "HURR", "SNOW", "TORN", "DROUGHT") # TODO: reorder these

state_map <- read_sf("outputs/state_estimates") # TODO: include estimates in next update (empty files for now)
state_map <- st_transform(state_map, "+proj=longlat +datum=WGS84")

cwa_map <- read_sf("outputs/cwa_estimates")
cwa_map <- cwa_map %>% 
  mutate_at(vars(all_of(vars_to_normalize)), list(~pnorm(.) * 100)) # convert z-scores to percentile ranks
cwa_map$CWA <- as.character(cwa_map$CWA)
cwa_map <- st_transform(cwa_map, "+proj=longlat +datum=WGS84")

fips_map <- read_sf("outputs/county_estimates")
fips_map <- fips_map %>% 
  mutate_at(vars(all_of(vars_to_normalize)), list(~pnorm(.) * 100)) # convert z-scores to percentile ranks
fips_map$CWA <- as.character(fips_map$CWA)
fips_map <- st_transform(fips_map, "+proj=longlat +datum=WGS84")

survey_data <- fread("outputs/base_survey_data.csv")

# Define demographic groups
survey_data$Gender <- factor(survey_data$MALE, labels = c("Female", "Male"))
survey_data$Ethnicity <- factor(survey_data$HISP, labels = c("Non-Hispanic", "Hispanic"))
survey_data$Race <- factor(survey_data$RACE_GROUP, labels = c("White", "Black or African American", "Other Race"))
survey_data$Age <- factor(survey_data$AGE_GROUP, labels = c("18-34", "35-59", "60+"))
survey_data <- survey_data %>% 
  mutate(Education = case_when(
    edu %in% 1:3 ~ "Less than College",
    edu %in% 4:6 ~ "Some College/College Degree",
    edu %in% 7:8 ~ "More than College"))
survey_data$Education <- factor(survey_data$Education, levels = c("Less than College", "Some College/College Degree", "More than College"))
survey_data$Region <- factor(survey_data$nws_region)
# survey_data$Numeracy <- factor(survey_data$Numeracy, levels = c("Low", "Moderate", "High")) # TODO: fix this in base survey file
survey_data$Area <- factor(survey_data$rural, labels = c("Urban", "Suburban", "Rural"))
survey_data <- survey_data %>% 
  mutate(Understanding = case_when(
    edu %in% 1 ~ "High",
    edu %in% 2 ~ "Moderate",
    edu %in% 3:5 ~ "Low"))
survey_data$Understanding <- factor(survey_data$Understanding, levels = c("Low", "Moderate", "High"))
survey_data$Year <- factor(survey_data$survey_year)
survey_data$Hazard <- factor(survey_data$survey_hazard, levels = c("WX", "TC"), labels = c("Severe", "Tropical"))

survey_data$All <- "All"
groups <- c("All" = "All", 
            "Gender" = "Gender", 
            "Ethnicity" = "Ethnicity", 
            "Race" = "Race", 
            "Age Group" = "Age",
            "NWS Region" = "Region", 
            "Education" = "Education", 
            # "Numeracy" = "Numeracy", # TODO: fix this in base survey file
            "Geographic Area (Urban/Rural)" = "Area", 
            "Understanding of Extreme Weather" = "Understanding", 
            "Survey Year" = "Year",
            "Survey Hazard" = "Hazard")

variable_data <- fread("variable_data_live.csv")
```

Composite Scales
======================================================================

Sidebar{.sidebar data-width=300}
-----------------------------------------------------------------------
<br/>
**Welcome!**<br/>
Welcome to WxDash, an interactive platform that provides access to data from the *Extreme Weather and Society Project*, which is run by researchers at the University of Oklahoma's [Institute for Public Policy Research and Analysis (IPPRA)](https://ou.edu/ippra). The data for the project come from yearly surveys that measure forecast and warning reception, comprehension, and response among adults (age 18+) that match the demographic characteristics of the US population. The statistics we present on this page are estimates that we calculate using a small area estimation technique called Multilevel Regression and Postratification (MRP). Navigate to the [About](https://crcm.shinyapps.io/WxDashExp/#section-about) tab to learn more about the project and the estimates. Click [here](https://dataverse.harvard.edu/dataverse/wxsurvey) to access the survey data.
<br/>
<br/>
**Instructions for this page.**<br/>
Select a measure from the drop down menu (above) to see an *average person percentile (APP) rank* estimate for each County Warning Area (CWA). These estimates compare the average percentile rank of adults who live in a CWA to the distribution of adults in the US. For example, a APP rank estimate of 62 indicates that, on average, adults who live in that CWA are above the national average; they score higher than 62% of adults across the country. If you click on a CWA, you will see the APP rank estimate and the percentile rank of the CWA comparison to other CWAs. If you click on *County Estimates*, the map will display APP estimates by county and the percentile rank of the county in comparison to other counties in the CWA.
<br/>
<br/>
**Questions or feedback?**<br/> 
Contact [Joe Ripberger](mailto:jtr@ou.edu) or [Makenzie Krocak](mailto:mjkrocak@ou.edu) at OU IPPRA.

Row {data-height=500}{.tabset .tabset-fade}
-----------------------------------------------------------------------

### Severe

```{r}
div(style = "font-size: 11px",
    selectInput(inputId = 'to_variable',
                label = 'Select composite measure to explore',
                choices = c('Tornado Warning Reception' = 'TO_RECEP',
                            'Tornado Warning Comprehension (Subjective)' = 'TO_SUB_COM',
                            'Tornado Warning Comprehension (Objective)' = 'TO_OBJ_COM',
                            'Tornado Warning Response' = 'TO_RESP',
                            'Efficacy of Protective Actions' = 'TO_EFFICAC'))
    )
```

```{r}
variable_map_cols <- viridis(100)
variable_map_pal <- colorNumeric(palette = variable_map_cols, domain = c(30, 70))

miniContentPanel(
  renderLeaflet({
  cwa_map$to_variable <- cwa_map %>% pull(input$to_variable)
  fips_map$to_variable <- fips_map %>% pull(input$to_variable)
  cwa_popup <- with(cwa_map, paste0("NWS ", City, ", ", ST, " (", CWA,")", "<br>",
                                    "Percentile rank of the average person in the CWA (APP): ", round(to_variable, 2), "<br>",
                                    "Percentile rank of the CWA: ", round(pnorm(scale(to_variable)) * 100, 2)))
  fips_popup <- with(fips_map, paste0(NAME, ", County ", "(", CWA, ")", "<br>",
                                      "Percentile rank of the average person in the County: ", round(to_variable, 2), "<br>",
                                      "Percentile rank of the County: ", round(pnorm(scale(to_variable)) * 100, 2)))
  map <- leaflet() %>%
    setView(lng = -97, lat = 40, zoom = 4) %>%
    addTiles() %>%
    addPolygons(
      data = cwa_map,
      fillColor = ~variable_map_pal(to_variable),
      highlight = highlightOptions(weight = 5, color = "grey", fillOpacity = 0.9, bringToFront = FALSE),
      color = "white",
      fillOpacity = 0.9,
      stroke = TRUE,
      smoothFactor = 0.8,
      weight = 2,
      popup = cwa_popup,
      layerId = ~CWA,
      group = "CWA Estimates") %>% 
    addPolygons(
      data = fips_map,
      fillColor = ~variable_map_pal(to_variable),
      highlight = highlightOptions(weight = 5, color = "grey", fillOpacity = 0.9, bringToFront = FALSE),
      color = "white",
      fillOpacity = 0.9,
      stroke = TRUE,
      smoothFactor = 0.8,
      weight = 0.5,
      popup = fips_popup,
      layerId = ~FIPS,
      group = "County Estimates") %>% 
    addPolylines(
      data = cwa_map,
      group = "CWA Borders",
      color = "white",
      weight = 3) %>%
    addPolylines(
      data = state_map,
      group = "State Borders",
      color = "#D3D3D3",
      weight = 4) %>%
    addLayersControl(
      baseGroups = c("CWA Estimates", "County Estimates"), # TODO: include state estimates in next update
      overlayGroups = c("State Borders", "CWA Borders"),
      options = layersControlOptions(collapsed = FALSE)) %>%
    showGroup("CWA") %>%
    # hideGroup("State") %>%
    hideGroup("County") %>% 
    hideGroup("CWA Borders") %>% 
    hideGroup("State Borders") %>% 
    addLegend(
      data = cwa_map,
      position = 'bottomright',
      pal = variable_map_pal,
      values = ~to_variable,
      opacity = 0.7,
      title = "Mean<br>Percentile<br>Rank")
  }), 
  padding = c(0, 0, 75, 0), scrollable = TRUE)
```

### Tropical

```{r}
div(style = "font-size: 11px",
    selectInput(inputId = 'hu_variable',
                label = 'Select composite measure to explore',
                choices = c('Hurricane Warning Reception' = 'HU_RECEP',
                            'Hurricane Warning Comprehension (Subjective)' = 'HU_SUB_COM',
                            'Hurricane Warning Comprehension (Objective)' = 'HU_OBJ_COM',
                            'Hurricane Warning Response' = 'HU_RESP'))
    )
```

```{r}
miniContentPanel(
  renderLeaflet({
  cwa_map$hu_variable <- cwa_map %>% pull(input$hu_variable)
  fips_map$hu_variable <- fips_map %>% pull(input$hu_variable)
  cwa_popup <- with(cwa_map, paste0("NWS ", City, ", ", ST, " (", CWA,")", "<br>",
                                    "Percentile rank of the average person in the CWA (APP): ", round(hu_variable, 2), "<br>",
                                    "Percentile rank of the CWA: ", round(pnorm(scale(hu_variable)) * 100, 2)))
  fips_popup <- with(fips_map, paste0(NAME, ", County ", "(", CWA, ")", "<br>",
                                      "Percentile rank of the average person in the County: ", round(hu_variable, 2), "<br>",
                                      "Percentile rank of the County: ", round(pnorm(scale(hu_variable)) * 100, 2)))
  leaflet() %>%
    setView(lng = -97, lat = 40, zoom = 4) %>%
    addTiles() %>%
    addPolygons(
      data = cwa_map,
      fillColor = ~variable_map_pal(hu_variable),
      highlight = highlightOptions(weight = 5, color = "grey", fillOpacity = 0.9, bringToFront = FALSE),
      color = "white",
      fillOpacity = 0.9,
      stroke = TRUE,
      smoothFactor = 0.8,
      weight = 2,
      popup = cwa_popup,
      layerId = ~CWA,
      group = "CWA Estimates") %>% 
    addPolygons(
      data = fips_map,
      fillColor = ~variable_map_pal(hu_variable),
      highlight = highlightOptions(weight = 5, color = "grey", fillOpacity = 0.9, bringToFront = FALSE),
      color = "white",
      fillOpacity = 0.9,
      stroke = TRUE,
      smoothFactor = 0.8,
      weight = 0.5,
      popup = fips_popup,
      layerId = ~FIPS,
      group = "County Estimates") %>% 
    addPolylines(
      data = cwa_map,
      group = "CWA Borders",
      color = "white",
      weight = 3) %>%
    addPolylines(
      data = state_map,
      group = "State Borders",
      color = "#D3D3D3",
      weight = 4) %>%
    addLayersControl(
      baseGroups = c("CWA Estimates", "County Estimates"), # TODO: include state estimates in next update
      overlayGroups = c("State Borders", "CWA Borders"),
      options = layersControlOptions(collapsed = FALSE)) %>%
    showGroup("CWA") %>%
    # hideGroup("State") %>%
    hideGroup("County") %>% 
    hideGroup("CWA Borders") %>% 
    hideGroup("State Borders") %>% 
    addLegend(
      data = cwa_map,
      position = 'bottomright',
      pal = variable_map_pal,
      values = ~hu_variable,
      opacity = 0.7,
      title = "Mean<br>Percentile<br>Rank")
  }), 
  padding = c(0, 0, 75, 0), scrollable = TRUE)
```

### Winter

Coming soon!

### WRN

Coming soon!

Risk Perceptions
======================================================================

Sidebar {.sidebar data-width=300}
-----------------------------------------------------------------------
<br/>
```{r}
selectInput(inputId = 'event', label = 'Select a type of event', choices = c(
  'Extreme Heat Waves' = 'PER_HEA',
  'Drought' = 'PER_DRO',
  'Extreme Cold Temperatures' = 'PER_COL',
  'Snow/Ice Storms' = 'PER_SNO',
  'Tornadoes' = 'PER_TOR',
  'Flooding' = 'PER_FLO',
  'Hurricanes' = 'PER_HUR',
  'Wildfires' = 'PER_FIR'),
  selected = "PER_DRO")
```
**Notes.**<br/>
Data on ***risk perceptions*** come from the following survey questions: Thinking about all four seasons (winter, summer, spring, and fall), how do you rate the risk of the following extreme weather events to you and the people in your area?

  - Extreme heat waves
  - Droughts
  - Extreme cold temperatures
  - Extreme snow (or ice) storms
  - Tornadoes
  - Floods
  - Hurricanes
  - Wildfires

Data on ***event frequency*** come from the [NOAA Storm Events Database](https://www.ncdc.noaa.gov/stormevents/) and the [US Drought Monitor](https://droughtmonitor.unl.edu).

Row {data-height=300}
-----------------------------------------------------------------------

### Subjective Risk Perceptions from Survey Data
```{r}
perception_map_cols <- viridis(100)
perception_map_pal <- colorNumeric(palette = perception_map_cols, domain = c(0, 100))

miniContentPanel(
  renderLeaflet({
  cwa_map$sub_rsk_variable <- cwa_map %>% pull(input$event)
  cwa_popup <- with(cwa_map, paste0("NWS ", City, ", ", ST, " (", CWA,")", "<br>",
                                    "Percentile rank of the average person in the CWA: ", round(sub_rsk_variable, 2), "<br>",
                                    "Percentile rank of the CWA: ", round(pnorm(scale(sub_rsk_variable)) * 100, 2)))
  leaflet() %>%
    setView(lng = -96.5, lat = 36.5, zoom = 3) %>%
    addTiles() %>%
    addPolygons(
      data = cwa_map,
      fillColor = ~perception_map_pal(round(pnorm(scale(sub_rsk_variable)) * 100, 2)),
      highlight = highlightOptions(weight = 5, color = "grey", fillOpacity = 0.9, bringToFront = FALSE),
      color = "white",
      fillOpacity = 0.9,
      stroke = TRUE,
      smoothFactor = 0.8,
      weight = 2,
      popup = cwa_popup,
      layerId = ~CWA,
      group = "CWA") %>%
    addLegendNumeric(
      orientation = c("horizontal"),
      data = cwa_map,
      position = 'bottomright',
      pal = perception_map_pal,
      values = ~round(pnorm(scale(sub_rsk_variable)) * 100, 2),
      fillOpacity = 0.7,
      height = 10, width = 200,
      title = "Percentile Rank of the CWA")
  }), 
  padding = c(0, 0, 0, 0), scrollable = TRUE)
```

### Objective Event Frequency from NOAA Storm Data
```{r}
frequency_map_cols <- viridis(100)
frequency_map_pal <- colorNumeric(palette = frequency_map_cols, domain = c(0, 100))

miniContentPanel(
  renderLeaflet({
  rsk_name <- NA
  rsk_name <- ifelse(input$event == "PER_HEA", "HEAT", rsk_name)
  rsk_name <- ifelse(input$event == "PER_DRO", "DROUGHT", rsk_name)
  rsk_name <- ifelse(input$event == "PER_COL", "COLD", rsk_name)
  rsk_name <- ifelse(input$event == "PER_SNO", "SNOW", rsk_name)
  rsk_name <- ifelse(input$event == "PER_TOR", "TORN", rsk_name)
  rsk_name <- ifelse(input$event == "PER_FLO", "FLOOD", rsk_name)
  rsk_name <- ifelse(input$event == "PER_HUR", "HURR", rsk_name)
  rsk_name <- ifelse(input$event == "PER_FIR", "FIRE", rsk_name)
  cwa_map$obj_frq_variable <- cwa_map %>% pull(rsk_name)
  cwa_popup <- with(cwa_map, paste0("NWS ", City, ", ", ST, " (", CWA,")", "<br>",
                                    "Percentile rank of the CWA: ", round(obj_frq_variable, 2)))
  leaflet() %>%
    setView(lng = -96.5, lat = 36.5, zoom = 3) %>%
    addTiles() %>%
    addPolygons(
      data = cwa_map,
      fillColor = ~frequency_map_pal(obj_frq_variable),
      highlight = highlightOptions(weight = 5, color = "grey", fillOpacity = 0.9, bringToFront = FALSE),
      color = "white",
      fillOpacity = 0.9,
      stroke = TRUE,
      smoothFactor = 0.8,
      weight = 2,
      popup = cwa_popup,
      layerId = ~CWA,
      group = "CWA") %>%
    addLegendNumeric(
      orientation = c("horizontal"),
      data = cwa_map,
      position = 'bottomright',
      pal = frequency_map_pal,
      values = ~obj_frq_variable,
      fillOpacity = 0.7,
      height = 10, width = 200,
      title = "Percentile Rank of the CWA")
  }), 
  padding = c(0, 0, 0, 0), scrollable = TRUE)
```

### Subjective Risk Perceptions vs. Objective Event Frequency
```{r}
difference_map_cols <- viridis(100, option = "cividis")
difference_map_pal <- colorNumeric(palette = difference_map_cols, domain = c(-75, 75))
cwa_map <- cwa_map %>%
  mutate(heat_difference = (pnorm(scale(cwa_map$PER_HEA)) * 100) - HEAT,
         drought_difference = (pnorm(scale(cwa_map$PER_DRO)) * 100) - DROUGHT,
         cold_difference = (pnorm(scale(cwa_map$PER_COL)) * 100) - COLD,
         snow_difference = (pnorm(scale(cwa_map$PER_SNO)) * 100) - SNOW,
         torn_difference = (pnorm(scale(cwa_map$PER_TOR)) * 100) - TORN,
         flood_difference = (pnorm(scale(cwa_map$PER_FLO)) * 100) - FLOOD,
         hurr_difference = (pnorm(scale(cwa_map$PER_HUR)) * 100) - HURR,
         fire_difference = (pnorm(scale(cwa_map$PER_FIR)) * 100) - FIRE)

miniContentPanel(
  renderLeaflet({
  dif_name <- NA
  dif_name <- ifelse(input$event == "PER_HEA", "heat_difference", dif_name)
  dif_name <- ifelse(input$event == "PER_DRO", "drought_difference", dif_name)
  dif_name <- ifelse(input$event == "PER_COL", "cold_difference", dif_name)
  dif_name <- ifelse(input$event == "PER_SNO", "snow_difference", dif_name)
  dif_name <- ifelse(input$event == "PER_TOR", "torn_difference", dif_name)
  dif_name <- ifelse(input$event == "PER_FLO", "flood_difference", dif_name)
  dif_name <- ifelse(input$event == "PER_HUR", "hurr_difference", dif_name)
  dif_name <- ifelse(input$event == "PER_FIR", "fire_difference", dif_name)
  cwa_map$sub_obj_variable <- cwa_map %>% pull(dif_name)
  cwa_popup <- with(cwa_map, paste0("NWS ", City, ", ", ST, " (", CWA,")", "<br>",
                                    "Difference in Percentile rank of the CWA: ", round(sub_obj_variable, 2)))
  leaflet() %>%
    setView(lng = -96.5, lat = 36.5, zoom = 3) %>%
    addTiles() %>%
    addPolygons(
      data = cwa_map,
      fillColor = ~difference_map_pal(sub_obj_variable),
      highlight = highlightOptions(weight = 5, color = "grey", fillOpacity = 0.9, bringToFront = FALSE),
      color = "white",
      fillOpacity = 0.9,
      stroke = TRUE,
      smoothFactor = 0.8,
      weight = 2,
      popup = cwa_popup,
      layerId = ~CWA,
      group = "CWA") %>%
    addLegendNumeric(
      orientation = c("horizontal"),
      data = cwa_map,
      position = 'bottomright',
      pal = difference_map_pal,
      values = ~sub_obj_variable,
      fillOpacity = 0.7,
      height = 10, width = 200,
      title = "Difference in Percentile Rank of the CWA")
  }), 
  padding = c(0, 0, 0, 0), scrollable = TRUE)
```

Row {data-height=400}
-----------------------------------------------------------------------
```{r}
lookup <- structure(c(as.character(cwa_map$CWA)), .Names = c(paste0("NWS ", cwa_map$City, ", ", cwa_map$ST, " (", cwa_map$CWA,")")))
lookup <- lookup[order(factor(names(lookup), levels = sort(names(lookup))))]
```

### Subjective Risk Perceptions vs. Objective Event Frequency by CWA
```{r}
selectInput(inputId = 'cwa', label = 'Select a CWA to highlight', choices = lookup)
cwa_map <- cwa_map %>%
  mutate(SCALE_PER_HEA = (pnorm(scale(cwa_map$PER_HEA)) * 100),
         SCALE_PER_DRO = (pnorm(scale(cwa_map$PER_DRO)) * 100),
         SCALE_PER_COL = (pnorm(scale(cwa_map$PER_COL)) * 100),
         SCALE_PER_SNO = (pnorm(scale(cwa_map$PER_SNO)) * 100),
         SCALE_PER_TOR = (pnorm(scale(cwa_map$PER_TOR)) * 100),
         SCALE_PER_FLO = (pnorm(scale(cwa_map$PER_FLO)) * 100),
         SCALE_PER_HUR = (pnorm(scale(cwa_map$PER_HUR)) * 100),
         SCALE_PER_FIR = (pnorm(scale(cwa_map$PER_FIR)) * 100))

miniContentPanel(renderPlot({
  sub_name <- NA
  sub_name <- ifelse(input$event == "PER_HEA", "SCALE_PER_HEA", sub_name)
  sub_name <- ifelse(input$event == "PER_DRO", "SCALE_PER_DRO", sub_name)
  sub_name <- ifelse(input$event == "PER_COL", "SCALE_PER_COL", sub_name)
  sub_name <- ifelse(input$event == "PER_SNO", "SCALE_PER_SNO", sub_name)
  sub_name <- ifelse(input$event == "PER_TOR", "SCALE_PER_TOR", sub_name)
  sub_name <- ifelse(input$event == "PER_FLO", "SCALE_PER_FLO", sub_name)
  sub_name <- ifelse(input$event == "PER_HUR", "SCALE_PER_HUR", sub_name)
  sub_name <- ifelse(input$event == "PER_FIR", "SCALE_PER_FIR", sub_name)
  obj_name <- NA
  obj_name <- ifelse(input$event == "PER_HEA", "HEAT", obj_name)
  obj_name <- ifelse(input$event == "PER_DRO", "DROUGHT", obj_name)
  obj_name <- ifelse(input$event == "PER_COL", "COLD", obj_name)
  obj_name <- ifelse(input$event == "PER_SNO", "SNOW", obj_name)
  obj_name <- ifelse(input$event == "PER_TOR", "TORN", obj_name)
  obj_name <- ifelse(input$event == "PER_FLO", "FLOOD", obj_name)
  obj_name <- ifelse(input$event == "PER_HUR", "HURR", obj_name)
  obj_name <- ifelse(input$event == "PER_FIR", "FIRE", obj_name)
  dif_name <- NA
  dif_name <- ifelse(input$event == "PER_HEA", "heat_difference", dif_name)
  dif_name <- ifelse(input$event == "PER_DRO", "drought_difference", dif_name)
  dif_name <- ifelse(input$event == "PER_COL", "cold_difference", dif_name)
  dif_name <- ifelse(input$event == "PER_SNO", "snow_difference", dif_name)
  dif_name <- ifelse(input$event == "PER_TOR", "torn_difference", dif_name)
  dif_name <- ifelse(input$event == "PER_FLO", "flood_difference", dif_name)
  dif_name <- ifelse(input$event == "PER_HUR", "hurr_difference", dif_name)
  dif_name <- ifelse(input$event == "PER_FIR", "fire_difference", dif_name)
  p <- ggplot(data = cwa_map, aes_string(x = obj_name, y = sub_name, color = dif_name)) +
    geom_smooth(method = lm, color = "red", alpha = 0.5, se = FALSE, size = 3) +
    geom_point(alpha = 0.3, size = 8) +
    geom_point(data = cwa_map[cwa_map$CWA == input$cwa, ],
               aes_string(x = obj_name, y = sub_name),
               color = "blue", fill = "blue", alpha = 0.3, size = 20) +
    geom_text(aes(label = CWA)) +
    labs(x = "Objective Event Frequency\n(Percentile Rank of the CWA)", y = "Subjective Risk Perceptions\n(Percentile Rank of the CWA)", title = input$cwa) +
    theme_classic() +
    theme(text = element_text(size = 20)) +
    guides(color = "none") +
    scale_color_viridis(option = "cividis") +
    lims(x = c(0, 100), y = c(0, 100))
  print(p)
  }), padding = c(0, 0, 75, 0), scrollable = TRUE)
```

Information Sources
======================================================================

Sidebar {.sidebar data-width=300}
-----------------------------------------------------------------------
<br/>
```{r}
selectInput(inputId = 'source_group', label = 'Select a grouping', choices = groups)
checkboxInput(inputId = 'check', label = 'Show text labels?', value = FALSE)
```
**Notes.**<br/>
Data on **information sources** come from the following survey questions:

  - How much do you, personally, rely on the following sources of information about [severe|tropical] weather?
  - Please indicate your level of trust in information about [severe|tropical] weather from each of the following organizations and groups.

Select the *Survey Hazard* grouping to see the difference across hazards.

Row
-----------------------------------------------------------------------

```{r}
rely_data <- reactive({
  survey_data %>%
    drop_na(!!as.name(input$source_group)) %>%
    group_by(!!as.name(input$source_group)) %>%
    select(wx_info1:wx_info8) %>%
    summarise_all(mean, na.rm = TRUE) %>%
    gather(variable, value, -!!as.name(input$source_group)) %>%
    mutate(Source = recode(variable,
                           "wx_info1" = "Broadcast\nRadio",
                           "wx_info2" = "Weather\nRadio",
                           "wx_info3" = "Television",
                           "wx_info8" = "Siren",
                           "wx_info4" = "Internet",
                           "wx_info5" = "Social\nMedia",
                           "wx_info6" = "Word\nof Mouth",
                           "wx_info7" = "Phone\nNotification")) %>%
    mutate(Source = factor(Source,
                        levels = c("Television", "Phone\nNotification", "Internet", "Siren", "Weather\nRadio",
                                   "Broadcast\nRadio", "Word\nof Mouth", "Social\nMedia"))) %>%
    mutate(l = round(value, 1))
  })

trust_data <- reactive({
  survey_data %>%
    drop_na(!!as.name(input$source_group)) %>%
    group_by(!!as.name(input$source_group)) %>%
    select(nws_trust:fam_trust) %>%
    summarise_all(mean, na.rm = TRUE) %>%
    gather(variable, value, -!!as.name(input$source_group)) %>%
    mutate(Source = recode(variable,
                           "nws_trust" = "National\nWeather\nService",
                           "lotv_trust" = "Regional\nor local TV\nstations",
                           "natv_trust" = "National TV\nstations",
                           "em_trust" = "State or\nlocal emergency\nmanagers",
                           "fam_trust" = "Family, friends,\nneighbors, employers,\nco-workers, etc.")) %>%
    mutate(Source = factor(Source,
                        levels = c("National\nWeather\nService",
                                   "Regional\nor local TV\nstations",
                                   "National TV\nstations",
                                   "State or\nlocal emergency\nmanagers",
                                   "Family, friends,\nneighbors, employers,\nco-workers, etc."))) %>%
    mutate(l = round(value, 1))
  })

renderPlot({
  p1 <- ggplot(rely_data(), aes_string(x = "Source", y = "value", color = input$source_group, group = input$source_group, label = "l")) +
    geom_point(size = 9, position = position_dodge(width = 0.7)) +
    theme_minimal() +
    theme(text = element_text(size = 18),
          plot.title = element_text(size = 18)) +
    labs(x = "", y = "Mean Response") +
    scale_y_continuous(breaks = 1:5, labels = c("Not\nmuch (1)", "Little (2)", "Somewhat (3)", "Much (4)", "A great\ndeal (5)"), limits = c(1, 5)) +
    ggtitle("How much do you rely on each of the following sources of information about extreme weather?") +
    theme(plot.title = element_text(color = "grey30")) +
    scale_color_viridis(discrete = TRUE) +
      theme(plot.title = element_textbox_simple(margin = margin(0, 0, 22, 0), fill = NA))
  p2 <- ggplot(trust_data(), aes_string(x = "Source", y = "value", color = input$source_group, group = input$source_group, label = "l")) +
    geom_point(size = 9, position = position_dodge(width = 0.7)) +
    theme_minimal() +
    theme(text = element_text(size = 18),
          plot.title = element_text(size = 18)) +
    labs(x = "", y = "Mean Response") +
    scale_y_continuous(breaks = 1:5, labels = c("No\nTrust (1)", "Low (2)", "Moderate (3)", "High (4)", "Complete\nTrust (5)"), limits = c(1, 5)) +
    ggtitle("Please indicate your level of trust in information about hurricanes from each of the following organizations and groups.") +
    theme(plot.title = element_text(color = "grey30")) +
    scale_color_viridis(discrete = TRUE) +
    theme(plot.title = element_textbox_simple(margin = margin(0, 0, 22, 0), fill = NA))
  p <- plot_grid(p1, p2, align = "v", ncol = 1)
  print(p)

  p1t <- p1 + geom_text(position = position_dodge(width = 0.7), size = 4, color = "white", fontface = "bold")
  p2t <- p2 + geom_text(position = position_dodge(width = 0.7), size = 4, color = "white", fontface = "bold")
  pt <- plot_grid(p1t, p2t, align = "v", ncol = 1)
  print(pt)

  if (input$check == FALSE) {print(p)}
  if (input$check == TRUE) {print(pt)}

  }, height = 900)
```

Explore Surveys
======================================================================

Sidebar {.sidebar data-width=300}
-----------------------------------------------------------------------
<br/>
```{r}
selectInput(inputId = 'exp_group', label = 'Select a grouping', choices = groups)
```
**Notes.**<br/>
Click on a survey question in the table to the right and a subgroup in the drop-down list above to see the distribution of survey responses to each question by population subgroup.

Row
-----------------------------------------------------------------------

### Distribution
```{r}
num <- survey_data %>%
  summarise_all(n_distinct) %>%
  select_if(any_vars((.) > 11))

cat <- survey_data %>%
  summarise_all(n_distinct) %>%
  select_if(any_vars((.) <= 11)) # all cat for now

d <- reactive({
  survey_data %>% 
    drop_na(weightfactor, input$exp_group, select_react() %>% pull(var = id)) %>% 
    as_survey_design(ids = 1, weights = weightfactor) %>% 
    group_by(!!as.name(input$exp_group), !!as.name(select_react() %>% pull(var = id))) %>%
    summarize(p = survey_mean(vartype = "ci")) %>% 
    mutate(p = round(p * 100, 0), p_low = p_low * 100, p_upp = p_upp * 100) %>% 
    mutate(l = paste0(p, "%")) %>% 
    mutate(resp = factor(!!as.name(select_react() %>% pull(var = id))))
  })

# renderTable({d()}) # use to see raw results in table

renderPlot({
  p1 <- ggplot(d() %>% drop_na(input$exp_group, select_react() %>% pull(var = id)),
               aes_string(x = "as.factor(resp)", y = "p", fill = input$exp_group, label = "l")) +
    geom_bar(stat = "identity", position = "dodge") +
    geom_text(position = position_dodge(width = .9), size = 5, vjust = -0.2, color = "black") +
    labs(x = "Response", y = "Respondents (%)", fill = "Group",
         title = variable_data[variable_data$id == select_react() %>% pull(var = id), ]$question,
         caption = paste0(
           "[",
           variable_data[variable_data$id == select_react() %>% pull(var = id), ]$response_options,
           "]")) +
    theme_classic(base_size = 20) +
    scale_fill_viridis(discrete = TRUE) +
    scale_y_continuous(minor_breaks = NULL,
                       expand = expansion(mult = c(0, 0.2))) +
    # scale_x_discrete(labels = variable_data[variable_data$id == select_react() %>% pull(var = id), ]$label) +
    theme(title = element_text(size = 14)) +
    theme(plot.title = element_textbox_simple(margin = margin(0, 0, 22, 0), fill = NA))

  p2 <- ggplot(survey_data %>% drop_na(input$exp_group, select_react() %>% pull(var = id)),
               aes_string(y = select_react() %>% pull(var = id), x = as.factor(input$exp_group), fill = input$exp_group)) +
    geom_boxplot() +
    labs(x = "Group", y = "Response", fill = "Group",
         title = variable_data[variable_data$id == select_react() %>% pull(var = id), ]$question,
         caption = paste0(
           "[",
           variable_data[variable_data$id == select_react() %>% pull(var = id), ]$response_options,
           "]")) +
    theme_classic(base_size = 20) +
    scale_fill_viridis(discrete = TRUE) +
    theme(title = element_text(size = 14)) +
    theme(plot.title = element_textbox_simple(margin = margin(0, 0, 22, 0), fill = NA))

  if (select_react() %>% pull(var = id) %in% names(cat)) {print(p1)}
  if (select_react() %>% pull(var = id) %in% names(num)) {print(p2)}
  })
```

Row
-----------------------------------------------------------------------

### Questions (click on a question)

```{r}
font.size <- "10pt"

datatable(variable_data %>%
            select(question, response_options),
          filter = "top",
          selection = "single",
          rownames = FALSE,
          style = "bootstrap",
          class = "compact",
          options = list(dom = "Blrtip",
                         initComplete = JS("function(settings, json) {",
                                           paste0("$(this.api().table().container()).css({'font-size': '",
                                                  font.size, "'});"),"}")), # governs text size
          colnames = c(
            "Question Text" = "question",
            "Response Options" = "response_options"),
          elementId = 'table')

variable_data <- variable_data %>% mutate(row_num = c(1:nrow(variable_data)))

select_react <- reactive({
  if(is.null(input$table_row_last_clicked)){
    variable_data %>%
      filter(row_num == 1) %>%
      select(id)}
  else {variable_data %>%
      filter(row_num == input$table_row_last_clicked) %>%
      select(id)}
  })
```

Time Series
======================================================================
Sidebar {.sidebar data-width=300}
-----------------------------------------------------------------------
<br/>
```{r}
ts_data <- survey_data %>%
  select(data.frame(groups)$groups, weightfactor, follow:risk_fire, alert_und, nws_trust:wx_info8, tor_map_und:svr_watchwarn_und, rec_all:rec_large_group, 
         rec_stream, resp_ignore:resp_large_group, resep_stream, wthr_info_paper:wthr_info_phone)
selectInput(inputId = 'ts_group', label = 'Select a grouping', choices = groups[ !groups == "Year"]) # remove survey year from list
# checkboxInput(inputId = 'scale', label = 'Show full y-scale?', value = FALSE) #TODO: fix this in a future version
```
**Notes.**<br/>
Click on a survey question in the table to the right and a subgroup in the drop-down list above to see the distribution of survey responses to each question by population subgroup.

Row
-----------------------------------------------------------------------

### Time Series Plot

```{r}
g <- reactive({
  ts_data %>%
    drop_na(weightfactor, input$ts_group, ts_select_react() %>% pull(var = id)) %>% 
    as_survey_design(ids = 1, weights = weightfactor) %>%
    group_by(Year, !!as.name(input$ts_group)) %>%
    summarize(mean = survey_mean(!!as.name(ts_select_react() %>% pull(var = id)), vartype = "ci")) %>% 
    mutate(mean = round(mean, 2))
})

# renderTable({g()}) # use to see raw results in table

renderPlot({
  # ylim <- c(min(ts_data[, ts_select_react() %>% pull(var = id)], na.rm = TRUE),
  #           max(ts_data[, ts_select_react() %>% pull(var = id)], na.rm = TRUE))
  p <- ggplot(g(), aes_string(x = "Year", y = "mean", color = input$ts_group, group = input$ts_group)) +
    geom_point(position = position_dodge(width = 0.1), size = 3) +
    geom_line(position = position_dodge(width = 0.1), size = 1) +
    geom_errorbar(aes(ymin = mean_low, ymax = mean_upp), position = position_dodge(width = 0.1), width = 0, size = 1) +
    labs(x = "Year", y = "Mean", fill = "Group",
         title = variable_data[variable_data$id == ts_select_react() %>% pull(var = id), ]$question,
                  caption = paste0(
           "[",
           variable_data[variable_data$id == ts_select_react() %>% pull(var = id), ]$response_options,
           "]")) +
    # scale_y_continuous(breaks = min(ylim):max(ylim), limits = ylim, expand = expansion(mult = c(0, 0))) +
    theme_classic(base_size = 20) +
    theme(title = element_text(size = 14)) +
    scale_color_viridis(discrete = TRUE) +
    theme(plot.title = element_textbox_simple(margin = margin(0, 0, 22, 0), fill = NA))
  print(p)
  # ps <- p + lims(y = c(min(ts_select_react() %>% pull(var = id), na.rm = TRUE), max(ts_select_react()() %>% pull(var = id), na.rm = TRUE)))
  # if (input$scale == FALSE) {print(p)}
  # if (input$scale == TRUE) {print(ps)}
})
```

Row
-----------------------------------------------------------------------

### Questions (click on a question)

```{r}
font.size <- "10pt"

ts_variable_data <- variable_data %>%
  filter(id %in% names(ts_data)) %>%
  mutate(row_num = c(1:nrow(.)))

datatable(ts_variable_data %>%
            select(question, response_options),
          filter = "top",
          selection = "single",
          rownames = FALSE,
          style = "bootstrap",
          class = "compact",
          options = list(dom = "Blrtip",
                         initComplete = JS("function(settings, json) {",
                                           paste0("$(this.api().table().container()).css({'font-size': '",
                                                  font.size, "'});"),"}")), # governs text size
          colnames = c(
            "Question Text" = "question",
            "Response Options" = "response_options"),
          elementId = 'table1')

ts_select_react <- reactive({
  if(is.null(input$table1_row_last_clicked)){
    ts_variable_data %>%
      filter(row_num == 1) %>%
      select(id)}
  else {ts_variable_data %>%
      filter(row_num == input$table1_row_last_clicked) %>%
      select(id)}
  })
```


Demographic Vulnerabilities
======================================================================

Sidebar{.sidebar data-width=200}
-----------------------------------------------------------------------
<br/>
**Notes.**<br/>
The CDC's Social Vulnerability Index (SVI) uses this set of 15 US census variables to help officials identify communities that may need support in preparing for hazards or recovering from disasters. Click [here](https://www.atsdr.cdc.gov/placeandhealth/svi/index.html) for more information.
<br/>

Row {data-height=500}{.tabset .tabset-fade}
-----------------------------------------------------------------------

### CDC SVI

```{r}
div(style = "font-size: 11px",
    selectInput(inputId = 'dem_variable', label = 'Select a demographic or theme', choices = c(
      'Percentage of persons below 150% poverty estimate' = 'POV',
      'Percentage of civilian (age 16+) unemployed' = 'UNEMP',
      'Percentage of housing cost-burdened occupied housing units with annual income less than $75,000' = 'HBURD',
      'Percentage of persons with no high school diploma (age 25+)' = 'NOHSDP',
      'Percentage uninsured in the total civilian noninstitutionalized population' = 'UNINSUR',
      'Percentage of persons aged 65 and older' = 'AGE65',
      'Percentage of persons aged 17 and younger' = 'AGE17',
      'Percentage of civilian noninstitutionalized population with a disability' = 'DISABL',
      'Percentage of single parent households with children under 18' = 'SNGPNT',
      'Percentage of persons (age 5+) who speak English less than well' = 'LIMENG',
      'Percentage minority (all persons except white, non-Hispanic)' = 'MINRTY',
      'Percentage of housing in structures with 10 or more units' = 'MUNIT',
      'Percentage of mobile homes' = 'MOBILE',
      'Percentage of occupied housing units with more people than rooms' = 'CROWD',
      'Percentage of households with no vehicle available' = 'NOVEH',
      'Percentage of persons in institutionalized group quarters' = 'GROUPQ',
      'Percentage of households without a computer with a broadband Internet subscription estimate' = 'NOINT',
      'Percentage of Black/African American, not Hispanic or Latino persons' = 'AFAM',
      'Percentage of Hispanic or Latino persons' = 'HISP',
      'Percentage of Asian, not Hispanic or Latino persons' = 'ASIAN',
      'Percentage of American Indian or Alaska Native, not Hispanic or Latino persons' = 'AIAN',
      'Percentage of Native Hawaiian or Other Pacific Islander, not Hispanic or Latino persons' = 'NHPI',
      'Percentage of two or more races, not Hispanic or Latino persons' = 'TWOMORE',
      'Percentage of some other race, not Hispanic or Latino persons' = 'OTHERRACE',
      'Percentage of persons in institutionalized group quarters' = 'THEME1',
      'Theme 2 (z-score): Household composition & disability (aged 65 or older, aged 17 or younger, older than age 5 with a disability, single-parent households)' = 'THEME2',
      'Theme 3 (z-score): Minority status & language (minority, speak English “less than well”)' = 'THEME3',
      'Theme 4 (z-score): Housing type & transportation (multi-unit structures, mobile homes, crowding, no vehicle, group quarters)' = 'THEME4',
      'Overall Social Vulnerability (z-score)' = 'THEMES'),
      width = "600px")
    )
```

```{r}
miniContentPanel(
  renderLeaflet({
  variable_map_pal <- colorNumeric(palette = viridis(100), domain = fips_map %>% pull(input$dem_variable))
  cwa_map$dem_variable <- cwa_map %>% pull(input$dem_variable)
  fips_map$dem_variable <- fips_map %>% pull(input$dem_variable)
  cwa_popup <- with(cwa_map, paste0("NWS ", City, ", ", ST, " (", CWA,")", "<br>",
                                    "Estimate: ", round(dem_variable, 2), "<br>",
                                    "Percentile rank of the CWA: ", round(pnorm(scale(dem_variable)) * 100, 2)))
  fips_popup <- with(fips_map, paste0(NAME, ", County ", "(", CWA, ")", "<br>",
                                      "Estimate: ", round(dem_variable, 2), "<br>",
                                      "Percentile rank of the County: ", round(pnorm(scale(dem_variable)) * 100, 2)))
  leaflet() %>%
    setView(lng = -97, lat = 40, zoom = 4) %>%
    addTiles() %>%
    addPolygons(
      data = cwa_map,
      fillColor = ~variable_map_pal(dem_variable),
      highlight = highlightOptions(weight = 5, color = "grey", fillOpacity = 0.9, bringToFront = FALSE),
      color = "white",
      fillOpacity = 0.9,
      stroke = TRUE,
      smoothFactor = 0.8,
      weight = 2,
      popup = cwa_popup,
      layerId = ~CWA,
      group = "CWA Estimates") %>%
    addPolygons(
      data = fips_map,
      fillColor = ~variable_map_pal(dem_variable),
      highlight = highlightOptions(weight = 5, color = "grey", fillOpacity = 0.9, bringToFront = FALSE),
      color = "white",
      fillOpacity = 0.9,
      stroke = TRUE,
      smoothFactor = 0.8,
      weight = 0.5,
      popup = fips_popup,
      layerId = ~FIPS,
      group = "County Estimates") %>%
    addPolylines(
      data = cwa_map,
      group = "CWA Borders",
      color = "white",
      weight = 3) %>%
    addPolylines(
      data = state_map,
      group = "State Borders",
      color = "#D3D3D3",
      weight = 4) %>%
    addLayersControl(
      baseGroups = c("CWA Estimates", "County Estimates"), # TODO: include state estimates in next update
      overlayGroups = c("State Borders", "CWA Borders"),
      options = layersControlOptions(collapsed = FALSE)) %>%
    showGroup("County Estimates") %>%
    hideGroup("CWA Borders") %>%
    # hideGroup("State") %>%
    hideGroup("CWA Estimates") %>%
    hideGroup("State Borders") %>%
    addLegend(
      data = fips_map,
      position = 'bottomright',
      pal = variable_map_pal,
      values = ~dem_variable,
      opacity = 0.7,
      title = "Estimate")
  }), 
  padding = c(0, 0, 75, 0), scrollable = TRUE)
```

### FEMA Risk Index

Coming soon!

About
======================================================================
**Motivation.** Members of the weather enterprise, including National Weather Service (NWS) forecasters, emergency managers, broadcast meteorologists, and private partners have many responsibilities, ranging from the issuance of forecasts and warnings during high impact weather events to outreach and public education campaigns during less turbulent periods. Effective education and risk communication across this range of responsibilities requires knowledge of the communities that enterprise members serve. This includes knowledge about atmospheric and climate conditions in communities as well as knowledge about the people in communities. Enterprise members often have access to a wide variety data that facilitate the first type of knowledge, but relatively little data on the populations they serve. As a result, it can be difficult to answer basic questions, such as what risks do the people in my community worry about or neglect? Do they generally receive, understand, and respond to forecasts and warnings? What sources of information do they rely on and trust? Absent reliable answers to these questions, it is challenging to develop public education and risk communication strategies that match the characteristics of the community.

**Methodology.** This project works to overcome some these challenges by developing a database of community statistics and this interactive platform that provides some of this information. The approach we use leverages population scale data from the Extreme Weather and Society Survey (Wx Survey) and known sub-population characteristics from the US Census to estimate statistics of interest to the weather enterprise. Run by the [University of Oklahoma Institute for Public Policy Research and Analysis (IPPRA)](https://ou.edu/ippra), the Wx Survey is a yearly survey of the US public that includes two types of questions: (1) baseline questions that measure core concepts such as risk perceptions, forecast and warning reception, comprehension, and response, hazard comprehension, and trust in information sources; and (2) one-time questions and experiments that address various topics, such as the impact of uncertainty and probabilistic information on risk judgments and protective action decision making. Visit our [data repository](https://dataverse.harvard.edu/dataverse/wxsurvey) for more information about the surveys. For more information about the composite measures we produce from these surveys and the steps we take to ensure that the measures are reliable, see:

- [Measuring Tornado Warning Reception, Comprehension, and Response in the United States](https://doi.org/10.1175/WCAS-D-19-0015.1)

Large population surveys such as the Wx Survey provide valuable information about the population as a whole, but [small area estimation (SEA)](https://www.nap.edu/read/13174/chapter/7) is necessary to identify differences across geographic sub-populations. This project uses multilevel regression and poststratification [MRP](https://journals.sagepub.com/doi/epub/10.1177/1478929919864773) to provide these estimates by NWS County Warning Area (CWA) and county. For more information about this methodology and how we apply it in this context, see:

- [Exploring Community Differences in Tornado Warning Reception, Comprehension, and Response Across the United States](https://doi.org/10.1175/BAMS-D-19-0064.1)
- [Geographic Distributions of Extreme Weather Risk Perceptions in the United States](https://doi.org/10.1111/risa.13569)

**Caution.** These data, like most survey data, come from non-probability samples that are not necessarily representative of the entire US population. Use cation when making inferences and drawing conclusions.

**Support.** Funding for this project comes from multiple sources. The University of Oklahoma provides support for data collection. The US Weather Program Office (WPO) and National Weather Service (NWS) provide support for  data analysis, programming, and maintenance.

**Data.** All of the survey data and metadata for the Extreme Weather and Society Project are available for download at the [Extreme Weather and Society Dataverse](https://dataverse.harvard.edu/dataverse/wxsurvey).

**Contact.** Please contact [Joe Ripberger](mailto:jtr@ou.edu) or [Makenzie Krocak](mailto:mjkrocak@ou.edu) if you have questions about the data or project.

<!-- HTML STYLE ----------------- -->
<style>

.section.sidebar {
  background-color: white;
  font-family: "Open-Sans", "Helvetica Neue", Helvetica, Arial, sans-serif;
  font-size: 13px;
}

.js-irs-0 .irs-bar {
  border-top-color: #443A83;
  border-bottom-color: #443A83;
}

.js-irs-0 .irs-bar-edge {
  border-color: #443A83;
}

.js-irs-0 .irs-single, .js-irs-0 .irs-bar-edge, .js-irs-0 .irs-bar {
  background: #443A83;
}

.navbar-inverse {
  background-color: #443A83;
  border-color: #440154;
  font-size: 15px;
}

.navbar-inverse .navbar-brand {
  color: #a3a9ac;
  font-size: 15px;
}

a:hover, a:focus {
color: #440154;
text-decoration: underline;
}

a {
color: #443A83;
text-decoration: none;
}

.navbar-inverse .navbar-nav>li>a {
  color: #a3a9ac;
  font-size: 15px;
}

</style>
